refactor: move from subsystems to protocols
